Refactored Event Store to have only async methods. Simplified the API structure.

+++ src/Core/Core/Objects/Versioning/IHaveVersion.cs
@@ -2,6 +2,6 @@
 {
     public interface IHaveVersion
     {
-        long Version { get; }
+        int Version { get; }
     }
 }
